git-svn-id: https://39.108.144.9:8443/svn/xhphp@216 23e8a065-ee1d-f14c-8958-478b19625c16

+++ projects/1.0.0/database/migrations/2016_07_24_205650_create_user_info_table.php
@@ -16,7 +16,7 @@ class CreateUserInfoTable extends Migration
             $table->increments('id');
             $table->integer('uid')->comment('用户ID');
             $table->tinyInteger('gender')->default(0)->comment('性别:0保密,1男,2女');
-            $table->tinyInteger('college_id')->default(0)->comment('所在学校ID');
+            $table->integer('college_id')->default(0)->comment('所在学校ID');
             $table->string('student_id')->default('')->comment('学号');
             $table->integer('enrollment_year')->comment('入学年份');
             $table->integer('birth_year')->comment('生日（年）');
